+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ArduinoJson: change log
 =======================
 
-HEAD
-----
+v5.0.8
+------
 
 * Made the library compatible with [PlatformIO](http://platformio.org/) (issue #181)
 * Fixed `JsonVariant::is<bool>()` that was incorrectly returning false (issue #214)